ver 0.0.9b
-Fixed some bugs(some crashes hopefully, selling bug)
-This project is now under the GNU General Public License.
-Updated to version 0.54
This release is only for updating for 0.54. I couldn't work on the project, so there isn't anything new. The channels are almost ready and included, but it is still buggy, so ignore it.
(but if you still want to try it you need separate exe for every channel and only one masterserver)
-Fixed some problems with the new packets.